cloudflare-tunnel-worker/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-cloudflare>=5.0.0,<6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
